float output not support, so changed percentage to int

+++ xv6-public-fresh/defs.h
@@ -72,7 +72,7 @@ char*           kalloc(void);
 void            kfree(char*);
 void            kinit1(void*, void*);
 void            kinit2(void*, void*);
-float           percentFreePages(void);
+int             percentFreePages(void);
 
 // kbd.c
 void            kbdintr(void);
